% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eff.singleFinder.R
\name{Eff.singleFinder}
\alias{Eff.singleFinder}
\title{Eff.singleFinder, to detect the Normalization efficiencies of single reference genes (RGs)}
\usage{
Eff.singleFinder(
  rawdata,
  num_RG,
  num_TG,
  num_group,
  num_sample,
  bio_rep,
  name_RG,
  name_TG,
  Dis.cq,
  gsg.Ratio,
  pv.th,
  Primer.E
)
}
\arguments{
\item{rawdata}{The qRT-PCR data matrix. The column-names and row-names are the sample names and gene symbols, respectively.
The row-names preceded by reference genes (RGs) and followed by target genes (TGs)}

\item{num_RG}{The number of reference genes (RGs) in your raw data matrix}

\item{num_TG}{The number of target genes (TGs) in your raw data matrix}

\item{num_group}{The number of experimental groups in raw data}

\item{num_sample}{The number of biological samples in your raw data matrix, is equal to Num_rep multiplied by num_group}

\item{bio_rep}{The number of biological replicate in all experimental groups}

\item{name_RG}{The gene names of your all reference genes (RGs)}

\item{name_TG}{The gene names of your all target genes (TGs)}

\item{Dis.cq}{The Dis.cq is the setting gap of Cq value, between the maximum value and second maximum,
or between the minimum and second minimum in the same experimental group. It is generally set to 2}

\item{gsg.Ratio}{The ratio of the gold standard group filtered from all NFs, is generally set to 0.01}

\item{pv.th}{The threshold for identifying positive and negative results between any two paired comparison groups, it is generally set to 0.05}

\item{Primer.E}{The amplification efficiency (E) of all primers, should be a vector here.
If you have not tested the primers' amplification efficiencies, please let Primer.E=1}
}
\value{

}
\description{
Eff.singleFinder, to detect the Normalization efficiencies of single reference genes (RGs)
}
\examples{
Eff.singleFinder(rawdata, num_RG=12, num_TG=20, num_sample=54, num_group=18, bio_rep=3,
                  name_RG, name_TG, Dis.cq=2, gsg.Ratio=0.01, pv.th=0.05, Primer.E)

}
\concept{}
